swapped JS controller file to plugin and updated Turbolinks function. it now works without having to refresh the page
